backend/ai-assistant/requirements.txt
psycopg2-binary==2.9.9
openai==1.54.3